refactor(chess): rearrange chess package
