+++ src/main/java/chess/ConsoleApplication.java
@@ -1,0 +1,12 @@
+package chess;
+
+import chess.controller.ConsoleChessController;
+
+public class ConsoleApplication {
+
+    public static void main(String[] args) {
+        ConsoleChessController consoleChessController = new ConsoleChessController();
+        consoleChessController.run();
+    }
+
+}
