--- src/main/resources/templates/index2.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>체스 게임 대기방</title>
-  <link rel="stylesheet" href="./src/index.css">
-</head>
-<body>
-<h1>♟️ 체스 게임 대기방♟</h1>
-<div id="game-entry">
-  <form enctype="text/plain" action="/game" method="get">
-    <label for="game-entry">방 번호를 입력하여 시작하기 (없을 시 새로 방 생성)</label>
-    <br>
-    <input type="text" name="game" value="1234">
-    <button type="submit">입장하기</button>
-  </form>
-</div>
-</body>
-</html>
