+++ src/test/java/chess/domain/piece/RookTest.java
@@ -51,11 +51,11 @@ public class RookTest {
         assertThat(actual).isFalse();
     }
 
-    @DisplayName("룩의 name은 Rook이다.")
+    @DisplayName("룩의 name은 rook이다.")
     @Test
     void getName() {
         String actual = new Rook(Color.BLACK, Position.of("a1")).getName();
-        String expected = "Rook";
+        String expected = "rook";
 
         assertThat(actual).isEqualTo(expected);
     }
